refactor: TestMemberController가 TestMemberService만 의존하도록 변경 (#1059)
* refactor: TestMemberController가 TestMemberService만 의존하도록 변경

* refactor: 미사용 메서드, 의존성 제거

* refactor: TestMemberService로 테스트용 메서드 이동

* refactor: 로컬, 개발 환경에서만 테스트용 컨트롤러, 서비스 빈 생성하도록 변경

* refactor: 로그 수정
