+++ src/main/java/com/gdschongik/gdsc/domain/member/api/TestMemberController.java
@@ -1,10 +1,11 @@
 package com.gdschongik.gdsc.domain.member.api;
 
-import com.gdschongik.gdsc.domain.member.application.AdminMemberService;
-import com.gdschongik.gdsc.domain.member.application.OnboardingMemberService;
+import static com.gdschongik.gdsc.global.common.constant.EnvironmentConstant.*;
+
 import com.gdschongik.gdsc.domain.member.application.TestMemberService;
 import com.gdschongik.gdsc.domain.member.dto.request.MemberTokenByGithubHandleRequest;
 import com.gdschongik.gdsc.domain.member.dto.response.MemberTokenResponse;
+import com.gdschongik.gdsc.global.annotation.ConditionalOnProfile;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -16,11 +17,10 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/test/members")
 @RequiredArgsConstructor
+@ConditionalOnProfile({LOCAL, DEV})
 public class TestMemberController {
 
     private final TestMemberService testMemberService;
-    private final OnboardingMemberService onboardingMemberService;
-    private final AdminMemberService adminMemberService;
 
     @Operation(summary = "게스트 회원 생성", description = "테스트용 API입니다. 깃허브 핸들명을 입력받아 임시 회원을 생성합니다.")
     @PostMapping
@@ -33,14 +33,14 @@ public class TestMemberController {
     @PostMapping("/token/github-handle")
     public ResponseEntity<MemberTokenResponse> createTemporaryTokenByGithubHandle(
             @Valid @RequestBody MemberTokenByGithubHandleRequest request) {
-        var response = onboardingMemberService.createTemporaryTokenByGithubHandle(request);
+        var response = testMemberService.createTemporaryTokenByGithubHandle(request);
         return ResponseEntity.ok().body(response);
     }
 
     @Operation(summary = "게스트로 강등", description = "테스트용 API입니다. 현재 멤버 역할을 게스트로 강등시키기 위해 사용합니다.")
     @PatchMapping("/demotion")
     public ResponseEntity<Void> demoteToGuest() {
-        adminMemberService.demoteToGuestAndRegularRequirementToUnsatisfied();
+        testMemberService.demoteToGuestAndRegularRequirementToUnsatisfied();
         return ResponseEntity.ok().build();
     }
 }
